implement DAO and CRUD operations for Student

+++ src/main/java/org/example/entity/Student.java
@@ -1,21 +1,21 @@
 package org.example.entity;
 
 public class Student {
-    private Long id;
+    private int id;
     private String name;
     private int age;
 
-    public Student(Long id, String name, int age) {
+    public Student(int id, String name, int age) {
         this.id = id;
         this.name = name;
         this.age = age;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
